changed layout for easier addition of screens
changed layout for easier addition of screens
