+++ .vs/SAC-LIFE/v17/DocumentLayout.json
@@ -1,0 +1,71 @@
+{
+  "Version": 1,
+  "WorkspaceRootPath": "C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\",
+  "Documents": [
+    {
+      "AbsoluteMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\App.js||{3B902123-F8A7-4915-9F01-361F908088D0}",
+      "RelativeMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|solutionrelative:SacStateApp\\App.js||{3B902123-F8A7-4915-9F01-361F908088D0}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\src\\screens\\HomeScreen.js||{3B902123-F8A7-4915-9F01-361F908088D0}",
+      "RelativeMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|solutionrelative:SacStateApp\\src\\screens\\HomeScreen.js||{3B902123-F8A7-4915-9F01-361F908088D0}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\src\\screens\\DetailsScreen.js||{3B902123-F8A7-4915-9F01-361F908088D0}",
+      "RelativeMoniker": "D:0:0:{A2FE74E1-B743-11D0-AE1A-00A0C90FFFC3}|\u003CMiscFiles\u003E|solutionrelative:SacStateApp\\src\\screens\\DetailsScreen.js||{3B902123-F8A7-4915-9F01-361F908088D0}"
+    }
+  ],
+  "DocumentGroupContainers": [
+    {
+      "Orientation": 0,
+      "VerticalTabListWidth": 256,
+      "DocumentGroups": [
+        {
+          "DockedWidth": 200,
+          "SelectedChildIndex": 2,
+          "Children": [
+            {
+              "$type": "Document",
+              "DocumentIndex": 2,
+              "Title": "DetailsScreen.js",
+              "DocumentMoniker": "C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\src\\screens\\DetailsScreen.js",
+              "RelativeDocumentMoniker": "SacStateApp\\src\\screens\\DetailsScreen.js",
+              "ToolTip": "C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\src\\screens\\DetailsScreen.js",
+              "RelativeToolTip": "SacStateApp\\src\\screens\\DetailsScreen.js",
+              "ViewState": "AQIAAAAAAAAAAAAAAADwvwsAAAAAAAAA",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.001646|",
+              "WhenOpened": "2024-10-17T16:17:29.37Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 1,
+              "Title": "HomeScreen.js",
+              "DocumentMoniker": "C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\src\\screens\\HomeScreen.js",
+              "RelativeDocumentMoniker": "SacStateApp\\src\\screens\\HomeScreen.js",
+              "ToolTip": "C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\src\\screens\\HomeScreen.js",
+              "RelativeToolTip": "SacStateApp\\src\\screens\\HomeScreen.js",
+              "ViewState": "AQIAAAAAAAAAAAAAAAAAABcAAAAEAAAA",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.001646|",
+              "WhenOpened": "2024-10-17T16:17:04.183Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 0,
+              "Title": "App.js",
+              "DocumentMoniker": "C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\App.js",
+              "RelativeDocumentMoniker": "SacStateApp\\App.js",
+              "ToolTip": "C:\\Users\\Darry\\Documents\\GitHub\\SAC-LIFE\\SacStateApp\\App.js",
+              "RelativeToolTip": "SacStateApp\\App.js",
+              "ViewState": "AQIAAAAAAAAAAAAAAAAAABMAAAABAAAA",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.001646|",
+              "WhenOpened": "2024-10-17T16:14:14.858Z",
+              "EditorCaption": ""
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
