+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,10 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\SacStateApp",
+    "\\SacStateApp\\src",
+    "\\SacStateApp\\src\\screens"
+  ],
+  "SelectedNode": "\\SacStateApp\\App.js",
+  "PreviewInSolutionExplorer": false
+}
