+++ SacStateApp/src/screens/DetailsScreen.js
@@ -1,0 +1,22 @@
+// screens/DetailsScreen.js
+import React from 'react';
+import { View, Text, StyleSheet } from 'react-native';
+
+const DetailsScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Details Screen</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default DetailsScreen;
