+++ SacStateApp/src/screens/HomeScreen.js
@@ -1,0 +1,27 @@
+// screens/Dashboard.js
+import React from 'react';
+import { View, Text, Button, StyleSheet } from 'react-native';
+
+const Dashboard = ({ navigation }) => {
+  return (
+    <View style={styles.container}>
+      <Text>yooooo</Text>
+      <Text>mamafdafdaf</Text>
+      <Button
+        title="Go to Details"
+        onPress={() => navigation.navigate('Details')}
+      />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default Dashboard;
